+++ build.gradle
@@ -4,7 +4,7 @@ plugins {
   id 'application'
 }
 application {
-  mainClass = 'com.jeonsaeyukjun.jeonsaeyukjunbe.report.crawling.HighTaxCompare' // 정확한 클래스 경로로 수정
+  mainClass = 'com.jeonsaeyukjun.jeonsaeyukjunbe.report.crawling.AddHighTaxCompare'
 }
 
 group 'com.jeonsaeyukjun'
@@ -24,20 +24,25 @@ targetCompatibility = '17'
 tasks.withType(JavaCompile) {
   options.encoding = 'UTF-8'
 }
+task runAddHighTaxCompare(type: JavaExec) {
+  mainClass = 'com.jeonsaeyukjun.jeonsaeyukjunbe.report.crawling.AddHighTaxCompare'  // mainClass 경로 수정
+  classpath = sourceSets.main.runtimeClasspath
+}
 
 dependencies {
   compileOnly('javax.servlet:javax.servlet-api:4.0.1')
-
-// JSoup for web scraping
-  implementation 'org.jsoup:jsoup:1.15.4'
+  //log4j
+  implementation 'org.apache.logging.log4j:log4j-api:2.19.0'
+  implementation 'org.apache.logging.log4j:log4j-core:2.19.0'
 
   // MySQL JDBC driver
   implementation 'mysql:mysql-connector-java:8.0.33'
 
-  // Jackson for JSON parsing (필요 시 사용)
-  implementation 'com.fasterxml.jackson.core:jackson-databind:2.14.1'
+  // Jackson for JSON parsing
+  implementation 'com.fasterxml.jackson.core:jackson-databind:2.17.2'
+  implementation 'com.fasterxml.jackson.core:jackson-core:2.17.2'
 
-  // 테스트 프레임워크 (필요 시 추가)
+  // 테스트 프레임ㅉ워크 (필요 시 추가)
   testImplementation 'org.junit.jupiter:junit-jupiter-api:5.8.1'
   testRuntimeOnly 'org.junit.jupiter:junit-jupiter-engine:5.8.1'
 
@@ -88,7 +93,7 @@ dependencies {
   implementation 'org.aspectj:aspectjweaver:1.9.20'
 
   // Selenium Java 라이브러리
-  implementation 'org.seleniumhq.selenium:selenium-java:4.4.0'
+  implementation 'org.seleniumhq.selenium:selenium-java:4.18.0'
 
   // WebDriver Manager (브라우저 드라이버 자동 관리)
   implementation 'io.github.bonigarcia:webdrivermanager:5.3.2'
